requirements.txt
fastapi[standard]
ipykernel
uvicorn
pydantic
SQLAlchemy
psycopg2-binary
streamlit
schedule
apscheduler
requests
sendgrid
pandas
matplotlib
seaborn
python-jose
passlib
bcrypt